+++ tests/stubs/helix/misc.scm
@@ -3,8 +3,12 @@
 
 (provide set-status!
          set-error!
-         enqueue-thread-local-callback-with-delay)
+         enqueue-thread-local-callback-with-delay
+         push-component!
+         pop-last-component-by-name!)
 
 (define (set-status! message) void)
 (define (set-error! message) void)
 (define (enqueue-thread-local-callback-with-delay delay thunk) void)
+(define (push-component! component) void)
+(define (pop-last-component-by-name! name) void)